+++ src/Exception/InvalidTopLevelType.php
@@ -1,8 +1,0 @@
-<?php
-declare(strict_types = 1);
-
-namespace Innmind\MediaType\Exception;
-
-class InvalidTopLevelType extends DomainException
-{
-}