+++ src/MediaType.php
@@ -3,13 +3,9 @@ declare(strict_types = 1);
 
 namespace Innmind\MediaType;
 
-use Innmind\MediaType\Exception\{
-    InvalidTopLevelType,
-    DomainException,
-};
 use Innmind\Immutable\{
+    Attempt,
     Sequence,
-    Set,
     Str,
     Maybe,
 };
@@ -22,52 +18,52 @@ final class MediaType
     /** @see https://tools.ietf.org/html/rfc6838#section-4.2 */
     private const FORMAT = '[A-Za-z0-9][A-Za-z0-9!#$&^_.-]{0,126}';
 
-    private string $topLevel;
-    private string $subType;
-    private string $suffix;
-    /** @var Sequence<Parameter> */
-    private Sequence $parameters;
-
-    /**
-     * @no-named-arguments
-     */
-    public function __construct(
-        string $topLevel,
-        string $subType,
-        string $suffix = '',
-        Parameter ...$parameters,
+    private function __construct(
+        private TopLevel $topLevel,
+        private string $subType,
+        private string $suffix,
+        /** @var Sequence<Parameter> */
+        private Sequence $parameters,
     ) {
-        if (!self::topLevels()->contains($topLevel)) {
-            throw new InvalidTopLevelType($topLevel);
-        }
-
-        $format = self::FORMAT;
-        $regex = "~^$format$~";
-
-        if (!Str::of($subType)->matches($regex)) {
-            throw new DomainException($subType);
-        }
-
-        if ($suffix !== '' && !Str::of($suffix)->matches($regex)) {
-            throw new DomainException($suffix);
-        }
-
-        $this->topLevel = $topLevel;
-        $this->subType = $subType;
-        $this->suffix = $suffix;
-        $this->parameters = Sequence::of(...$parameters);
     }
 
     /**
      * @psalm-pure
-     * @throws DomainException
+     * @no-named-arguments
+     */
+    public static function from(
+        TopLevel $topLevel,
+        string $subType,
+        string $suffix = '',
+        Parameter ...$parameters,
+    ): self {
+        $format = self::FORMAT;
+        $regex = "~^$format$~";
+
+        if (!Str::of($subType)->matches($regex)) {
+            throw new \DomainException($subType);
+        }
+
+        if ($suffix !== '' && !Str::of($suffix)->matches($regex)) {
+            throw new \DomainException($suffix);
+        }
+
+        return new self(
+            $topLevel,
+            $subType,
+            $suffix,
+            Sequence::of(...$parameters),
+        );
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @throws \DomainException
      */
     public static function of(string $string): self
     {
-        return self::maybe($string)->match(
-            static fn($self) => $self,
-            static fn() => throw new DomainException($string),
-        );
+        return self::attempt($string)->unwrap();
     }
 
     /**
@@ -81,15 +77,32 @@ final class MediaType
             ->filter(static fn($string) => $string->matches(self::pattern()))
             ->map(static fn($string) => $string->pregSplit('~[;,] ?~'))
             ->flatMap(
-                static fn($splits) => self::capture($splits->first())->flatMap(
-                    static fn(Str $topLevel, Str $subType, Str $suffix) => self::build(
-                        $topLevel->toString(),
-                        $subType->toString(),
-                        $suffix->toString(),
-                        $splits->drop(1),
+                static fn($splits) => $splits
+                    ->first()
+                    ->flatMap(self::capture(...))
+                    ->flatMap(
+                        static fn($self) => $splits
+                            ->drop(1)
+                            ->map(static fn($parameter) => $parameter->toString())
+                            ->map(Parameter::of(...))
+                            ->sink($self)
+                            ->maybe(static fn($self, $parameter) => $parameter->map(
+                                $self->withParameter(...),
+                            )),
                     ),
-                ),
             );
+    }
+
+    /**
+     * @psalm-pure
+     *
+     * @return Attempt<self>
+     */
+    public static function attempt(string $string): Attempt
+    {
+        return self::maybe($string)->attempt(
+            static fn() => throw new \DomainException($string),
+        );
     }
 
     /**
@@ -97,10 +110,15 @@ final class MediaType
      */
     public static function null(): self
     {
-        return new self('application', 'octet-stream');
+        return new self(
+            TopLevel::application,
+            'octet-stream',
+            '',
+            Sequence::of(),
+        );
     }
 
-    public function topLevel(): string
+    public function topLevel(): TopLevel
     {
         return $this->topLevel;
     }
@@ -132,31 +150,20 @@ final class MediaType
 
         return \sprintf(
             '%s/%s%s%s',
-            $this->topLevel,
+            $this->topLevel->name,
             $this->subType,
             $this->suffix !== '' ? '+'.$this->suffix : '',
             !$parameters->empty() ? '; '.$parameters->toString() : '',
         );
     }
 
-    /**
-     * List of allowed top levels
-     *
-     * @return Set<string>
-     */
-    public static function topLevels(): Set
+    private function withParameter(Parameter $parameter): self
     {
-        return Set::strings(
-            'application',
-            'audio',
-            'font',
-            'example',
-            'image',
-            'message',
-            'model',
-            'multipart',
-            'text',
-            'video',
+        return new self(
+            $this->topLevel,
+            $this->subType,
+            $this->suffix,
+            ($this->parameters)($parameter),
         );
     }
 
@@ -166,68 +173,43 @@ final class MediaType
 
         return \sprintf(
             "~%s/$format(\+$format)?([;,] $format=[\w\-.]+)?~",
-            Str::of('|')->join(self::topLevels())->toString(),
+            Str::of('|')
+                ->join(
+                    Sequence::of(...TopLevel::cases())
+                        ->map(static fn($level) => $level->name),
+                )
+                ->toString(),
         );
     }
 
     /**
-     * @param Sequence<Str> $parameters
-     *
      * @return Maybe<self>
      */
-    private static function build(
-        string $topLevel,
-        string $subType,
-        string $suffix,
-        Sequence $parameters,
-    ): Maybe {
-        if ($parameters->empty()) {
-            return Maybe::just(new self($topLevel, $subType, $suffix));
-        }
-
-        /** @psalm-suppress NamedArgumentNotAllowed */
-        return self::captureParameters($parameters)->map(
-            static fn(Parameter ...$parameters) => new self(
-                $topLevel,
-                $subType,
-                $suffix,
-                ...$parameters,
-            ),
-        );
-    }
-
-    /**
-     * @param Maybe<Str> $string
-     */
-    private static function capture(Maybe $string): Maybe\Comprehension
+    private static function capture(Str $string): Maybe
     {
         $format = self::FORMAT;
+        $matches = $string->capture(\sprintf(
+            "~^(?<topLevel>%s)/(?<subType>$format)(\+(?<suffix>$format))?$~",
+            Str::of('|')
+                ->join(
+                    Sequence::of(...TopLevel::cases())
+                        ->map(static fn($level) => $level->name),
+                )
+                ->toString(),
+        ));
 
-        return $string
-            ->map(static fn($string) => $string->capture(\sprintf(
-                "~^(?<topLevel>%s)/(?<subType>$format)(\+(?<suffix>$format))?$~",
-                Str::of('|')->join(self::topLevels())->toString(),
-            )))
-            ->match(
-                static fn($matches) => Maybe::all(
-                    $matches->get('topLevel'),
-                    $matches->get('subType'),
-                    $matches->get('suffix')->otherwise(static fn() => Maybe::just(Str::of(''))),
-                ),
-                static fn() => Maybe::all(Maybe::nothing()),
-            );
-    }
-
-    /**
-     * @param Sequence<Str> $parameters
-     */
-    private static function captureParameters(Sequence $parameters): Maybe\Comprehension
-    {
-        return $parameters
-            ->map(static fn($parameter) => Parameter::of($parameter->toString()))
-            ->match(
-                static fn($first, $rest) => Maybe::all($first, ...$rest->toList()),
-                static fn() => Maybe::all(Maybe::nothing()),
-            );
+        return Maybe::all(
+            $matches
+                ->get('topLevel')
+                ->map(static fn($level) => $level->toString())
+                ->flatMap(TopLevel::maybe(...)),
+            $matches->get('subType'),
+            $matches->get('suffix')->otherwise(static fn() => Maybe::just(Str::of(''))),
+        )->map(static fn(TopLevel $topLevel, Str $subType, Str $suffix) => new self(
+            $topLevel,
+            $subType->toString(),
+            $suffix->toString(),
+            Sequence::of(),
+        ));
     }
 }