+++ src/Parameter.php
@@ -3,7 +3,6 @@ declare(strict_types = 1);
 
 namespace Innmind\MediaType;
 
-use Innmind\MediaType\Exception\DomainException;
 use Innmind\Immutable\{
     Str,
     Maybe,
@@ -22,16 +21,24 @@ final class Parameter
     private string $name;
     private string $value;
 
-    public function __construct(string $name, string $value)
+    private function __construct(string $name, string $value)
     {
         $format = self::NAME;
 
         if (!Str::of($name)->matches("~^$format$~")) {
-            throw new DomainException($name);
+            throw new \DomainException($name);
         }
 
         $this->name = $name;
         $this->value = $value;
+    }
+
+    /**
+     * @psalm-pure
+     */
+    public static function from(string $name, string $value): self
+    {
+        return new self($name, $value);
     }
 
     /**